= kamon-annotation: almost complete source code documentation
